fixed formatting

+++ app/src/main/kotlin/bad/rhythm/game/clone/Section.kt
@@ -1,6 +1,3 @@
 package bad.rhythm.game.clone
 
-class Section(var bpm: Int, var offset: Int, var hitObjects: Array<HitObject>) {
-
-
-}
+class Section(var bpm: Int, var offset: Double, var hitObjects: Array<HitObject>) {}
